+++ PHP/deleteGroup.php
@@ -1,9 +1,13 @@
 <?php
-	/*Connect*/
-	mysql_connect("62.113.225.192","web90","maer89");
-
-	/*select database*/
-	mysql_select_db("usr_web90_3");
+	include 'DB_data.php';
+	
+	// Connect with DB
+	$con = mysql_connect($ServerAdr, $UserName, $pw);
+	if(!$con) {
+		die('Cold not connect: ' .mysql_error());
+	}
+	
+	mysql_select_db($database,$con);
 	
 	$userid = $_POST["uid"];
 	$name = $_POST["name"];
@@ -12,8 +16,16 @@
 	//delete group
 	$res = mysql_query("DELETE FROM groups WHERE userID = $userid AND name = '".$name."'");
 	
+	if(!$res){
+			die('Error: ' . mysql_error());
+	}
+		
 	//delete messages
 	$res = mysql_query("DELETE FROM messages WHERE messageGroup = '".$name."' AND userID = $userid");
+	
+	if(!$res){
+			die('Error: ' . mysql_error());
+	}
 	
 	//delete answers
 	$res = mysql_query("SELECT id FROM messages WHERE messageGroup = '".$name."' AND userID = $userid");
